# swagger.yaml
openapi: 3.0.0
info:
  title: Planriean Subjects Service API
  version: '2.0.1'
  description: API for managing university course planning and registration
servers:
  - url: http://localhost:3030
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /plan-restrict/view/{restgrp_id}/subject:
    get:
      summary: Get restricted plan subjects
      security:
        - bearerAuth: []
      tags: [Plan Restricted]
      parameters:
        - name: restgrp_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restricted plan subjects retrieved
    put:
      summary: Update restricted plan subjects
      security:
        - bearerAuth: []
      tags: [Plan Restricted]
      parameters:
        - name: restgrp_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restricted plan subjects updated

  # Register Routes
...
security:
  - bearerAuth: []